Separates scgiUtil from clientRequestManager
